phonebooth V2/pyproject.toml
[project]
name = "phonebooth-backend"
version = "0.1.0"
description = "Discord-style backend with FastAPI + Supabase Postgres/Auth"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "asyncpg>=0.29.0",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "httpx>=0.27.0",
  "redis>=5.0.7",
  "discord.py>=2.4.0",
]

[tool.uv]
package = false
